Refactored PlaylistOperations with commands and all
Completes https://soundcloud.atlassian.net/browse/DROID-747

Squashed commit of the following:

commit 10837c4a8281b8bb26b073627bfeff0fc57cd04d
Author: jonschmidt <jon@soundcloud.com>
Date:   Thu Feb 26 16:53:05 2015 +0100

    More cleanups

    - convert more ID ints to longs
    - equalsVerifier on PlaylistInfo

commit 6887006af31d08ec235dbafacea3c1cfee6bfd4a
Author: jonschmidt <jon@soundcloud.com>
Date:   Thu Feb 26 10:48:33 2015 +0100

    cleanups

commit 114593fb017d02b5ebf350a336502f0f1c2fb52d
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 17:56:12 2015 +0100

    add permalink to ApiPlaylist storage

commit 11d04722f017a8b7c3290d3b1ff28ad08e0fee2f
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 17:54:07 2015 +0100

    fix return result in test

commit 71d44239bad059c7ec83631832c96836fa967f87
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 17:49:07 2015 +0100

    Cleanups

commit 4ab300e509d57d4264857bf2c0f7de5cfe4f6524
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 17:35:31 2015 +0100

    Enable new playlist sync feature flag in debug

commit 0dfb5cf10ee63df19ede9d1a04a2e3d04f0fda23
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 17:23:40 2015 +0100

    add feature flag for new synced to SyncInitiator

commit 00e459de76d509c23954aa96ff05ce6fc37ee6c6
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 14:07:04 2015 +0100

    Use longs to fix local IDs not being preserved

commit b13950ff0216c2e50019e135af72920a2b551126
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 12:22:40 2015 +0100

    local playlist considerations

commit 5d03d8fd4fcb3eb8e7784a5108be903eefaa98fd
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 11:30:26 2015 +0100

    Basic PlaylistOperations tests for load/sync

    This does not include local-only playlists yet

commit 4e8d2c56f9b4e70a4b6d4ca7143125018945749f
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 11:28:40 2015 +0100

    PermalinkUrls are null sometimes (have to investigate)

commit ee8a7613126b80406c5e645e2c61151b2a295ef9
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 11:28:08 2015 +0100

    Change default playlist track count (easier testing)

commit 3ef7d75d5e2f07a2866de96591fbd5d076cdd909
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 11:27:34 2015 +0100

    LoadPlaylistTracks command + tests

commit e31a3d4e1cc4d4b821b24dd5bef04c400c7262c3
Author: jonschmidt <jon@soundcloud.com>
Date:   Wed Feb 25 11:26:31 2015 +0100

    Command Test for empty missing playlist

+++ app/src/main/java/com/soundcloud/android/likes/UpdateLikeCommand.java
@@ -93,7 +93,7 @@ class UpdateLikeCommand extends Command<PropertySet, PropertySet, UpdateLikeComm
         public PropertySet map(CursorReader cursorReader) {
             final PropertySet propertySet = PropertySet.create(cursorReader.getColumnCount());
 
-            propertySet.put(TrackProperty.URN, Urn.forTrack(cursorReader.getInt(BaseColumns._ID)));
+            propertySet.put(TrackProperty.URN, Urn.forTrack(cursorReader.getLong(BaseColumns._ID)));
             propertySet.put(PlayableProperty.LIKES_COUNT, cursorReader.getInt(TableColumns.SoundView.LIKES_COUNT));
 
             return propertySet;
